Materials and event graph changes
This commit has some material and event graph changes made as part of my task to import different skins into the game whenever the user would like. while the functionality for importing different skins on the fly has not been implemented, there are no issues with performance and stability as a result.