Killing Spree Datafield Setup
Setup basic data fields for killing spree data to be tracked. Will need to work more on calculating shutdowns etc, but for now the player's current killstreak can be calculated, and their max streak will be recorded. Will need to decide on how many kills it takes to consider something a proper 'streak'