Killing Spree Calculations Update
Took awhile finding a way to grab the victim's playerstate in order to check if they were on a spree or not. Now have shutdowns a.k.a revenges being calculated, as well as the minute a player is considered on a killing spree their spree count will be incremented. Still have their max spree length being calculated upon death.