Debugging changes and basic structure for death messages
These changes put in place the basic structure for deciding which death message to output. Currently I have it set to display messages for debugging when the player is killed by various projectiles. Once the death message arrays are populated, the debugging information can be swapped out for the creative text.