Death Message implementation
Death messages have been implemented for when the player is killed by their own attacks. These death messages are pulled at random from the 'suicideMessage' array and will typically start with the player name. It should be noted that at present the 'playerName' is just the name of the player controller.