Decision structure and location change for death messages
The decision structure for the 'DeathMessages' function has been implemented, now a death message will be played for self kills, opponent kills, and stage related deaths. I have also changed the location of where the function is located, this is due to the information and data present in this location. Additionally, it should be noted that there is mention of displaying death messages in the 'RegisterKill' function, however upon close inspection I have found that these "messages" do not display to the screen and appear to have more to do with logs.