Coverage Test Inclusion
This is the first draft of a quick coverage test to test each of the fields I have implemented.  It is currently not being expressly called anywhere as it is designed to be called if/when devs see fit to test this functionality on the drop of a dime. This coverage test WILL be called locally on my machine during my debugging efforts, but a function call will not be committed to the repo.