Coverage test framework for death messages
I have created a function for the death message coverage tests and have given it a basic structure which I can continue to build upon in future commits. It should be noted that the pointer variables for this function have yet to be defined, I am currently looking into how to go about doing this for the given objects.