Optimize Test Coverage
Briefly went in to optimize the logic for the test coverage function being utilized to print more efficiently i.e not constantly printing streaks when testing multikills prints (technically a triple kill is also a kill streak, this can be seen in the first screenshot of test coverage). Also included a brief addendum that all multikills above 6 are still 'monster kills'.