Changes to coverage test function and minor bug fix
I have set 'CoverageTestDM' to now take in the same parameters as the 'DeathMessage' function, this will help to avoid errors and complexity caused by attempting to initialize pointers to AControllers. Additionally, I have added some code to the 'DeathMessages' function to handle the killer and victim pointers if they are null. Aside from that I have also formatted the output of the coverage test to be a bit more readable.