minor adjustments for 'DeathMessages' and 'CoverageTestDM'
I have made a few minor adjustments to the functions mentioned, such as removing a parameter that was not being used and commenting out the coverage test function call. I have also added a note on lines 446 - 450 where I mention how to activate and use the test coverage for death messages.