Doxyfile Addition
Included doxyfile with all relevant information, as well as a project logo for doxygen to use, the logo is of course curtesy of the original developers. Doxyfile paths should be fine, even though they look to be absolute paths, with other settings in the doxyfile one should be able to just pull this repo onto their machine, and run the doxyfile no problem without any changes to it.